+++ golem-rib/src/type_checker/mod.rs
@@ -1,6 +1,13 @@
-pub(crate) use call_args::*;
 pub(crate) use validate::*;
 pub(crate) use type_check_error::*;
-mod call_args;
-mod validate;
+pub(crate) use unresolved_types::*;
+pub(crate) use type_mismatch::*;
+pub(crate) use validate::*;
+pub(crate) use phases::*;
+
 mod type_check_error;
+mod unresolved_types;
+mod type_mismatch;
+mod validate;
+mod phases;
+
