+++ golem-rib/src/type_checker/phases/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) use call_args::*;
+mod call_args;
